utils: add utils/api.hh and utils/api.cc to share common methods from api
Some functions and classes defined in api/ may be reused in other parts
of Scylla. Their code was moved to utils/api.{cc,hh}.
