+++ configure.py
@@ -685,6 +685,7 @@ scylla_core = (['message/messaging_service.cc',
                 'row_cache.cc',
                 'schema_mutations.cc',
                 'generic_server.cc',
+                'utils/api.cc',
                 'utils/array-search.cc',
                 'utils/base64.cc',
                 'utils/logalloc.cc',
